Start "fresh"
There's so much crap to fix that we'll just start with nothing loaded
and add each module as they're fixed (if they can be).

A lot of things will be dropped due to changes in Dragonflight, such
as things blizzard added to replace things I've done, or made
incompatible with my stuff.